pyproject.toml
[project]
name = "sc2bot"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "python-sc2>=0.11.2",
]

[tool.black]
line-length = 100
